re-org composition; add more introspection; add `tup` node method
re-org composition files

re-org part II

re-org part III; add `tup` and more tests
