+++ src/composition/composites/pipeline_static.jl
