+++ test/composition/composites/from_network.jl
@@ -1,4 +1,4 @@
-module TestComposites
+module TestFromNetwork
 
 using Test
 using MLJBase
@@ -19,104 +19,6 @@ ytrain = yin[train]
 
 ridge_model = FooBarRegressor(lambda=0.1)
 selector_model = FeatureSelector()
-
-@testset "first test of hand-exported network" begin
-    composite = SimpleDeterministicCompositeModel(model=ridge_model,
-                                                  transformer=selector_model)
-
-    fitresult, cache, report = MLJBase.fit(composite, 3, Xtrain, ytrain)
-
-    # to check internals:
-    ridge = MLJBase.machines(fitresult)[1]
-    selector = MLJBase.machines(fitresult)[2]
-    ridge_old = deepcopy(ridge)
-    selector_old = deepcopy(selector)
-
-    # this should trigger no retraining:
-    fitresult, cache, report =
-        @test_logs(
-            (:info, r"^Not"),
-            (:info, r"^Not"),
-            MLJBase.update(composite, 2, fitresult, cache, Xtrain, ytrain))
-    @test ridge.fitresult == ridge_old.fitresult
-    @test selector.fitresult == selector_old.fitresult
-
-    # this should trigger update of selector and training of ridge:
-    selector_model.features = [:a, :b]
-    fitresult, cache, report =
-        @test_logs(
-            (:info, r"^Updating"),
-            (:info, r"^Training"),
-            MLJBase.update(composite, 2, fitresult, cache, Xtrain, ytrain))
-    @test ridge.fitresult != ridge_old.fitresult
-    @test selector.fitresult != selector_old.fitresult
-    ridge_old = deepcopy(ridge)
-    selector_old = deepcopy(selector)
-
-    # this should trigger updating of ridge only:
-    ridge_model.lambda = 1.0
-    fitresult, cache, report =
-        @test_logs(
-            (:info, r"^Not"),
-            (:info, r"^Updating"),
-            MLJBase.update(composite, 2, fitresult, cache, Xtrain, ytrain))
-    @test ridge.fitresult != ridge_old.fitresult
-    @test selector.fitresult == selector_old.fitresult
-
-    predict(composite, fitresult, MLJBase.selectrows(Xin, test))
-
-    Xs = source(Xtrain)
-    ys = source(ytrain, kind=:target)
-
-    mach = machine(composite, Xs, ys)
-    yhat = predict(mach, Xs)
-    fit!(yhat, verbosity=3)
-    composite.transformer.features = [:b, :c]
-    fit!(yhat, verbosity=3)
-    fit!(yhat, rows=1:20, verbosity=3)
-    yhat(MLJBase.selectrows(Xin, test))
-
-end
-
-mutable struct WrappedRidge <: DeterministicNetwork
-    ridge
-end
-
-@testset "second test of hand-exported network" begin
-
-    function MLJBase.fit(model::WrappedRidge, verbosity::Integer, X, y)
-        Xs = source(X)
-        ys = source(y, kind=:target)
-
-        stand = Standardizer()
-        standM = machine(stand, Xs)
-        W = transform(standM, Xs)
-
-        boxcox = UnivariateBoxCoxTransformer()
-        boxcoxM = machine(boxcox, ys)
-        z = transform(boxcoxM, ys)
-
-        ridgeM = machine(model.ridge, W, z)
-        zhat = predict(ridgeM, W)
-        yhat = inverse_transform(boxcoxM, zhat)
-
-        fit!(yhat)
-        return fitresults(yhat)
-    end
-
-    MLJBase.input_scitype(::Type{<:WrappedRidge}) = Table(Continuous)
-    MLJBase.target_scitype(::Type{<:WrappedRidge}) = AbstractVector{<:Continuous}
-
-    ridge = FooBarRegressor(lambda=0.1)
-    model_ = WrappedRidge(ridge)
-    mach = machine(model_, Xin, yin)
-    fit!(mach)
-    yhat=predict(mach, Xin)
-    ridge.lambda = 1.0
-    fit!(mach)
-    @test predict(mach, Xin) != yhat
-
-end
 
 @load DecisionTreeRegressor
 @load DecisionTreeClassifier
@@ -143,7 +45,7 @@ yhat = exp(zhat)
 
 ex = Meta.parse("Composite(knn_rgs=knn, one_hot_enc=hot) <= yhat")
 modeltype_ex, fieldname_exs, model_exs, N_ex, kind, trait_dic =
-    MLJBase.from_network_preprocess(TestComposites, ex)
+    MLJBase.from_network_preprocess(TestFromNetwork, ex)
 @test modeltype_ex == :Composite
 @test fieldname_exs == [:knn_rgs, :one_hot_enc]
 @test model_exs == [:knn, :hot]
@@ -160,13 +62,13 @@ zhat = inverse_transform(standM, uhat)
 yhat = exp(zhat)
 ex = Meta.parse("Composite(knn_rgs=knn, one_hot_enc=hot) <= yhat")
 modeltype_ex, fieldname_exs, model_exs, N_ex, kind, trait_dic =
-    MLJBase.from_network_preprocess(TestComposites, ex)
+    MLJBase.from_network_preprocess(TestFromNetwork, ex)
 @test trait_dic[:supports_weights]
 
 # unsupervised:
 ex = Meta.parse("Composite(one_hot_enc=hot) <= W")
 modeltype_ex, fieldname_exs, model_exs, N_ex, kind, trait_dic =
-    MLJBase.from_network_preprocess(TestComposites, ex)
+    MLJBase.from_network_preprocess(TestFromNetwork, ex)
 @test modeltype_ex == :Composite
 @test fieldname_exs == [:one_hot_enc,]
 @test model_exs == [:hot,]
@@ -186,7 +88,7 @@ yhat = predict(elmM, H)
 
 ex = Meta.parse("Composite(selector=fea,one_hot=hot,tree=elm) <= yhat")
 modeltype_ex, fieldname_exs, model_exs, N_ex, kind, trait_dic =
-    MLJBase.from_network_preprocess(TestComposites,
+    MLJBase.from_network_preprocess(TestFromNetwork,
                                 ex, :(prediction_type=:probabilistic))
 @test modeltype_ex == :Composite
 @test fieldname_exs == [:selector, :one_hot, :tree]
@@ -198,54 +100,54 @@ modeltype_ex, fieldname_exs, model_exs, N_ex, kind, trait_dic =
 
 ex = Meta.parse("45")
 @test_throws(ArgumentError,
-             MLJBase.from_network_preprocess(TestComposites,
+             MLJBase.from_network_preprocess(TestFromNetwork,
                                          ex, :(prediction_type=:probabilistic)))
 
 ex = Meta.parse("Composite(elm=elm) << yhat")
 @test_throws(ArgumentError,
-             MLJBase.from_network_preprocess(TestComposites,
+             MLJBase.from_network_preprocess(TestFromNetwork,
                                          ex, :(prediction_type=:probabilistic)))
 
 ex = Meta.parse("45")
 @test_throws(ArgumentError,
-             MLJBase.from_network_preprocess(TestComposites,
+             MLJBase.from_network_preprocess(TestFromNetwork,
                                          ex, :(prediction_type=:probabilistic)))
 
 ex = Meta.parse("45 <= yhat")
 @test_throws(ArgumentError,
-             MLJBase.from_network_preprocess(TestComposites,
+             MLJBase.from_network_preprocess(TestFromNetwork,
                                          ex, :(prediction_type=:probabilistic)))
 ex = Meta.parse("Comp(elm=45) <= yhat")
 @test_throws(ArgumentError,
-             MLJBase.from_network_preprocess(TestComposites,
+             MLJBase.from_network_preprocess(TestFromNetwork,
                                          ex, :(prediction_type=:probabilistic)))
 
 ex = Meta.parse("Comp(elm=>elm) <= yhat")
 @test_throws(ArgumentError,
-             MLJBase.from_network_preprocess(TestComposites,
+             MLJBase.from_network_preprocess(TestFromNetwork,
                                          ex, :(prediction_type=:probabilistic)))
 
 ex = Meta.parse("Comp(34=elm) <= yhat")
 @test_throws(ArgumentError,
-             MLJBase.from_network_preprocess(TestComposites,
+             MLJBase.from_network_preprocess(TestFromNetwork,
                                          ex, :(prediction_type=:probabilistic)))
 
 ex = Meta.parse("Comp(elm=elm) <= 45")
 @test_throws(ArgumentError,
-             MLJBase.from_network_preprocess(TestComposites,
+             MLJBase.from_network_preprocess(TestFromNetwork,
                                          ex, :(prediction_type=:probabilistic)))
 
 z = vcat(ys, ys)
 ex = Meta.parse("Comp() <= z")
 @test_throws(ArgumentError,
-             MLJBase.from_network_preprocess(TestComposites,ex))
+             MLJBase.from_network_preprocess(TestFromNetwork,ex))
 
 X2s = source(nothing)
 # z = @test_logs (:warn, r"^A node ref") vcat(Xs, X2s)
 z = vcat(Xs, X2s)
 ex = Meta.parse("Comp() <= z")
 @test_throws(ArgumentError,
-             MLJBase.from_network_preprocess(TestComposites, ex))
+             MLJBase.from_network_preprocess(TestFromNetwork, ex))
 
 
 y2s = source(nothing, kind=:target)
@@ -255,11 +157,11 @@ z = vcat(ys, y2s, Xs)
 
 ex = Meta.parse("Comp() <= z")
 @test_throws(ArgumentError,
-             MLJBase.from_network_preprocess(TestComposites, ex))
+             MLJBase.from_network_preprocess(TestFromNetwork, ex))
 
 ex = Meta.parse("Composite(one_hot_enc=hot) <= W")
 @test_throws(ArgumentError,
-             MLJBase.from_network_preprocess(TestComposites,
+             MLJBase.from_network_preprocess(TestFromNetwork,
                                          ex, :(prediction_type=:probabilistic)))
 
 
@@ -439,6 +341,7 @@ posterior = predict(mach, rows=1:div(N,2))[1]
 
 composite_with_no_fields = @from_network CompositeWithNoFields() <= yhat
 mach = fit!(machine(composite_with_no_fields, X, y))
+
 
 end
 true
