+++ test/composition/composites/pipeline_static.jl
