+++ test/composition/composites/pipelines.jl
