+++ test/composition/learning_networks/arrows.jl
