+++ test/composition/learning_networks/inspection.jl
@@ -1,0 +1,65 @@
+module TestLearingNetworksInspection
+
+using Test
+using MLJBase
+using ..Models
+
+@load KNNRegressor
+
+@constant X = source()
+@constant y = source(kind=:target)
+
+hot = OneHotEncoder()
+hotM = machine(hot, X)
+@constant W = transform(hotM, X)
+knn = KNNRegressor()
+knnM = machine(knn, W, y)
+@constant yhat = predict(knnM, W)
+@constant K =  2*X
+@constant all = tup(yhat, K)
+
+@test MLJBase.tree(yhat) == (operation = predict,
+                             model = knn,
+                             arg1 = (operation = transform,
+                                     model = hot,
+                                     arg1 = (source = X, ),
+                                     train_arg1 = (source = X, )),
+                             train_arg1 = (operation = transform,
+                                           model = hot,
+                                           arg1 = (source = X, ),
+                                           train_arg1 = (source = X, )),
+                             train_arg2 = (source = y,))
+
+@test Set(models(yhat)) == Set([hot, knn])
+@test Set(sources(yhat)) == Set([X, y])
+@test Set(origins(yhat)) == Set([X,])
+@test Set(machines(yhat)) == Set([knnM, hotM])
+@test Set(MLJBase.args(yhat)) == Set([W, ])
+@test Set(MLJBase.train_args(yhat)) == Set([W, y])
+@test Set(MLJBase.children(X, all)) == Set([W, K])
+
+@constant Q = 2X
+@constant R = 3X
+@constant S = tup(X, Q, R)
+@test Set(MLJBase.children(X, S)) == Set([Q, R, S])
+@test MLJBase.lower_bound([Int, Float64]) == Union{}
+@test MLJBase.lower_bound([Int, Integer]) == Int
+@test MLJBase.lower_bound([Int, Integer]) == Int
+@test MLJBase.lower_bound([]) == Any
+
+@test input_scitype(2X) == Unknown
+@test input_scitype(yhat) == input_scitype(KNNRegressor())
+W2 = transform(machine(UnivariateStandardizer(), X), X)
+@test input_scitype(X, tup(W, W2)) == Union{}
+@test input_scitype(X, tup(Q, W)) == Unknown
+
+y1 = predict(machine(@load(DecisionTreeRegressor), X, y), X)
+@test input_scitype(y1) == Table(Continuous, OrderedFactor, Count)
+y2 = predict(machine(@load(KNNRegressor), X, y), X)
+@test input_scitype(y2) == Table(Continuous)
+@test input_scitype(X, tup(y1, y2)) == Table(Continuous)
+@test input_scitype(X, tup(y1, y2, Q)) == Unknown
+
+end
+
+true
